refactor: rename footer to main-footer
